Travis will now test this branch
Travis will now test the correct branch

Solved potential Travis issue

More Travis fixes

Travis will now just test Hadoop GeoTiffRDD test

Solved another issue with Travis

Fixed catalog_test issue on Travis

Certain tests will now be skipped on Travis

Claend up catalog_test

Travis will now use more memory when running tests

Cleaned up more files

Travis will not cache pip

Possible fix for Travis failures

Skipping another test on Travis

Updated the catalog test

Trying another method of making Travis work

Tried something else to fix Travis

Removed files from geotiff_test_files

Updated s3 and hadoop geotiff rdd tests

Updated tests to make them more memory efficient

Made the tests a bit faster

More performance fixes for the tests

Fixed typo

Renamed the schema tests so that they have schema in the name

Travis will now run tests in batches

Fixed typo in .travis.yml

Fixed issue

Moved schema tests to tests/schema_tests/

Only files within the top directory will be collected

Made tests even faster

Added the one tile_layer_metadata_test back in
