+++ geopyspark/tests/costdistance_test.py
@@ -1,7 +1,8 @@
 import os
 import unittest
-import rasterio
 import numpy as np
+
+import pytest
 
 from shapely.geometry import Point
 from geopyspark.tests.base_test_class import BaseTestClass
@@ -10,8 +11,6 @@ from geopyspark.geotrellis.constants import SPATIAL
 
 
 class CostDistanceTest(BaseTestClass):
-    geopysc = BaseTestClass.geopysc
-
     data = np.array([[
         [1.0, 1.0, 1.0, 1.0, 1.0],
         [1.0, 1.0, 1.0, 1.0, 1.0],
@@ -23,7 +22,7 @@ class CostDistanceTest(BaseTestClass):
              ({'row': 1, 'col': 0}, {'no_data_value': -1.0, 'data': data}),
              ({'row': 0, 'col': 1}, {'no_data_value': -1.0, 'data': data}),
              ({'row': 1, 'col': 1}, {'no_data_value': -1.0, 'data': data})]
-    rdd = geopysc.pysc.parallelize(layer)
+    rdd = BaseTestClass.geopysc.pysc.parallelize(layer)
 
     extent = {'xmin': 0.0, 'ymin': 0.0, 'xmax': 33.0, 'ymax': 33.0}
     layout = {'layoutCols': 2, 'layoutRows': 2, 'tileCols': 5, 'tileRows': 5}
@@ -38,6 +37,11 @@ class CostDistanceTest(BaseTestClass):
                     'tileLayout': {'tileCols': 5, 'tileRows': 5, 'layoutCols': 2, 'layoutRows': 2}}}
 
     raster_rdd = TiledRasterRDD.from_numpy_rdd(BaseTestClass.geopysc, SPATIAL, rdd, metadata)
+
+    @pytest.fixture(autouse=True)
+    def tearDown(self):
+        yield
+        BaseTestClass.geopysc.pysc._gateway.close()
 
     def test_costdistance_finite(self):
         def zero_one(kv):
