+++ geopyspark/tests/focal_test.py
@@ -1,7 +1,7 @@
 import os
 import unittest
-import rasterio
 import numpy as np
+import pytest
 
 from geopyspark.geotrellis.rdd import TiledRasterRDD
 from geopyspark.tests.base_test_class import BaseTestClass
@@ -9,8 +9,6 @@ from geopyspark.geotrellis.constants import SPATIAL, SUM, MIN, SQUARE, ANNULUS
 
 
 class FocalTest(BaseTestClass):
-    geopysc = BaseTestClass.geopysc
-
     data = np.array([[
         [1.0, 1.0, 1.0, 1.0, 1.0],
         [1.0, 1.0, 1.0, 1.0, 1.0],
@@ -22,7 +20,7 @@ class FocalTest(BaseTestClass):
              ({'row': 1, 'col': 0}, {'no_data_value': -1.0, 'data': data}),
              ({'row': 0, 'col': 1}, {'no_data_value': -1.0, 'data': data}),
              ({'row': 1, 'col': 1}, {'no_data_value': -1.0, 'data': data})]
-    rdd = geopysc.pysc.parallelize(layer)
+    rdd = BaseTestClass.geopysc.pysc.parallelize(layer)
 
     extent = {'xmin': 0.0, 'ymin': 0.0, 'xmax': 33.0, 'ymax': 33.0}
     layout = {'layoutCols': 2, 'layoutRows': 2, 'tileCols': 5, 'tileRows': 5}
@@ -37,6 +35,11 @@ class FocalTest(BaseTestClass):
                     'tileLayout': {'tileCols': 5, 'tileRows': 5, 'layoutCols': 2, 'layoutRows': 2}}}
 
     raster_rdd = TiledRasterRDD.from_numpy_rdd(BaseTestClass.geopysc, SPATIAL, rdd, metadata)
+
+    @pytest.fixture(autouse=True)
+    def tearDown(self):
+        yield
+        BaseTestClass.geopysc.pysc._gateway.close()
 
     def test_focal_sum(self):
         result = self.raster_rdd.focal(
@@ -54,3 +57,4 @@ class FocalTest(BaseTestClass):
 
 if __name__ == "__main__":
     unittest.main()
+    BaseTestClass.geopysc.pysc.stop()
