+++ geopyspark/tests/hadoop_geotiff_rdd_test.py
@@ -1,6 +1,7 @@
 import unittest
 from os import walk, path
 import rasterio
+import pytest
 
 from geopyspark.geotrellis.constants import SPATIAL
 from geopyspark.tests.python_test_utils import check_directory, geotiff_test_path
@@ -45,6 +46,11 @@ class GeoTiffIOTest(object):
 class Singleband(GeoTiffIOTest, BaseTestClass):
     dir_path = geotiff_test_path("one-month-tiles/")
 
+    @pytest.fixture(autouse=True)
+    def tearDown(self):
+        yield
+        BaseTestClass.geopysc.pysc._gateway.close()
+
     def read_singleband_geotrellis(self, options=None):
         if options is None:
             result = get(BaseTestClass.geopysc,
@@ -69,7 +75,7 @@ class Singleband(GeoTiffIOTest, BaseTestClass):
             self.assertEqual(x['no_data_value'], y['no_data_value'])
 
     def windowed_result_checker(self, windowed_tiles):
-        self.assertEqual(len(windowed_tiles), 24)
+        self.assertEqual(len(windowed_tiles), 4)
 
     def test_windowed_tiles(self):
         geotrellis_tiles = self.read_singleband_geotrellis(True)
@@ -83,49 +89,5 @@ class Singleband(GeoTiffIOTest, BaseTestClass):
             self.assertTrue((x['data'] == y['data']).all())
             self.assertEqual(x['no_data_value'], y['no_data_value'])
 
-
-class Multiband(GeoTiffIOTest, BaseTestClass):
-    dir_path = geotiff_test_path("one-month-tiles-multiband/")
-    options = {'maxTileSize': 256}
-
-    def read_multiband_geotrellis(self, options=None):
-        if options is None:
-            result = get(BaseTestClass.geopysc,
-                         SPATIAL,
-                         self.dir_path)
-        else:
-            result = get(BaseTestClass.geopysc,
-                         SPATIAL,
-                         self.dir_path,
-                         options)
-
-        return [tile[1] for tile in result.to_numpy_rdd().collect()]
-
-    def test_whole_tiles(self):
-        geotrellis_tiles = self.read_multiband_geotrellis()
-
-        file_paths = self.get_filepaths(self.dir_path)
-        rasterio_tiles = self.read_geotiff_rasterio(file_paths, False)
-
-        for x, y in zip(geotrellis_tiles, rasterio_tiles):
-            self.assertTrue((x['data'] == y['data']).all())
-
-    def windowed_result_checker(self, windowed_tiles):
-        self.assertEqual(len(windowed_tiles), 4)
-
-    def test_windowed_tiles(self):
-
-        geotrellis_tiles = self.read_multiband_geotrellis(options=self.options)
-
-        file_paths = self.get_filepaths(self.dir_path)
-        rasterio_tiles = self.read_geotiff_rasterio(file_paths, True)
-
-        self.windowed_result_checker(geotrellis_tiles)
-
-        for x, y in zip(geotrellis_tiles, rasterio_tiles):
-            self.assertTrue((x['data'] == y['data']).all())
-
-
 if __name__ == "__main__":
-
     unittest.main()
