+++ geopyspark/tests/stitch_test.py
@@ -1,7 +1,7 @@
 import os
 import unittest
-import rasterio
 import numpy as np
+import pytest
 
 from shapely.geometry import Point
 from geopyspark.geotrellis.rdd import TiledRasterRDD
@@ -10,8 +10,6 @@ from geopyspark.geotrellis.constants import SPATIAL
 
 
 class StitchTest(BaseTestClass):
-    geopysc = BaseTestClass.geopysc
-
     data = np.array([[
         [1.0, 1.0, 1.0, 1.0, 1.0],
         [1.0, 1.0, 1.0, 1.0, 1.0],
@@ -23,7 +21,7 @@ class StitchTest(BaseTestClass):
              ({'row': 1, 'col': 0}, {'no_data_value': -1.0, 'data': data}),
              ({'row': 0, 'col': 1}, {'no_data_value': -1.0, 'data': data}),
              ({'row': 1, 'col': 1}, {'no_data_value': -1.0, 'data': data})]
-    rdd = geopysc.pysc.parallelize(layer)
+    rdd = BaseTestClass.geopysc.pysc.parallelize(layer)
 
     extent = {'xmin': 0.0, 'ymin': 0.0, 'xmax': 33.0, 'ymax': 33.0}
     layout = {'layoutCols': 2, 'layoutRows': 2, 'tileCols': 5, 'tileRows': 5}
@@ -38,6 +36,11 @@ class StitchTest(BaseTestClass):
                     'tileLayout': {'tileCols': 5, 'tileRows': 5, 'layoutCols': 2, 'layoutRows': 2}}}
 
     raster_rdd = TiledRasterRDD.from_numpy_rdd(BaseTestClass.geopysc, SPATIAL, rdd, metadata)
+
+    @pytest.fixture(scope='class', autouse=True)
+    def tearDown(self):
+        yield
+        BaseTestClass.geopysc.pysc._gateway.close()
 
     def test_stitch(self):
         result = self.raster_rdd.stitch()
