--- geopyspark/tests/temporal_projected_extent_test.py
@@ -1,44 +1,0 @@
-import unittest
-
-from pyspark import RDD
-from pyspark.serializers import AutoBatchedSerializer
-from geopyspark.avroserializer import AvroSerializer
-from geopyspark.tests.base_test_class import BaseTestClass
-from py4j.java_gateway import java_import
-
-
-class TemporalProjectedExtentSchemaTest(BaseTestClass):
-    path = "geopyspark.geotrellis.tests.schemas.TemporalProjectedExtentWrapper"
-    java_import(BaseTestClass.geopysc.pysc._gateway.jvm, path)
-
-    expected_tpextents = [
-        {'epsg': 2004, 'extent': {'xmax': 1.0, 'xmin': 0.0, 'ymax': 1.0, 'ymin': 0.0}, 'instant': 0, 'proj4': None},
-        {'epsg': 2004, 'extent': {'xmax': 3.0, 'xmin': 1.0, 'ymax': 4.0, 'ymin': 2.0}, 'instant': 1, 'proj4': None},
-        {'epsg': 2004, 'extent': {'xmax': 7.0, 'xmin': 5.0, 'ymax': 8.0, 'ymin': 6.0}, 'instant': 2, 'proj4': None}]
-
-    sc = BaseTestClass.geopysc.pysc._jsc.sc()
-    ew = BaseTestClass.geopysc.pysc._gateway.jvm.TemporalProjectedExtentWrapper
-
-    tup = ew.testOut(sc)
-    java_rdd = tup._1()
-    ser = AvroSerializer(tup._2())
-
-    rdd = RDD(java_rdd, BaseTestClass.geopysc.pysc, AutoBatchedSerializer(ser))
-    collected = rdd.collect()
-
-    def result_checker(self, actual_tpe, expected_tpe):
-        for actual, expected in zip(actual_tpe, expected_tpe):
-            self.assertDictEqual(actual, expected)
-
-    def test_encoded_tpextents(self):
-        encoded = self.rdd.map(lambda s: s)
-        actual_encoded = encoded.collect()
-
-        self.result_checker(actual_encoded, self.expected_tpextents)
-
-    def test_decoded_tpextents(self):
-        self.result_checker(self.collected, self.expected_tpextents)
-
-
-if __name__ == "__main__":
-    unittest.main()
